requirements.txt
argparse
numpy
pyyaml
simplejson
mpld3
flask
astropy